--- a/server/src/render-web/index.html
+++ b/server/src/render-web/index.html
@@ -30,7 +30,7 @@
       }</script>
   </head>
   <body>
-    <div id="root"></div>
+    <div id="root"><!-- APP_HERE --></div>
     <script src="/dist/bundle.js"></script>
   </body>
 </html>