--- a/web/src/components/pages.tsx
+++ b/web/src/components/pages.tsx
@@ -29,8 +29,6 @@ import Privacy from './pages/privacy';
 import Terms from './pages/terms';
 import NotFound from './pages/not-found';
 
-const showDatasetsPage = localStorage.getItem('showDatasetsPage');
-
 const shareURL = 'https://voice.mozilla.org/';
 const encodedShareText = encodeURIComponent(
   'Help robots talk, donate your voice at ' + shareURL
@@ -413,11 +411,9 @@ class Pages extends React.Component<PagesProps, PagesState> {
         <NavLink to={URLS.RECORD} exact>
           Speak
         </NavLink>
-        {showDatasetsPage && (
-          <NavLink to={URLS.DATA} exact>
-            Datasets
-          </NavLink>
-        )}
+        <NavLink to={URLS.DATA} exact>
+          Datasets
+        </NavLink>
         <NavLink to={URLS.PROFILE} exact>
           Profile
         </NavLink>