--- a/web/src/components/pages/home/home.css
+++ b/web/src/components/pages/home/home.css
@@ -38,7 +38,6 @@
 #home-title {
     position: absolute;
     top: 0;
-    animation: fade-in .5s;
 }
 
 #contribute-button {