--- a/web/src/components/vars.css
+++ b/web/src/components/vars.css
@@ -20,7 +20,6 @@
     --hero-height-mobile: 12.5rem;
     --robot-height: 22rem;
     --robot-height-mobile: 16rem;
-    --spinner-height: 10rem;
     --easing: cubic-bezier(0.4, 0.0, 0.2, 1);
     --shadow-overhang: 0 1px 0 rgba(12,13,14,0.1),
     0 1px 3px rgba(12,13,14,0.1),