--- a/web/webpack.dev.config.js
+++ b/web/webpack.dev.config.js
@@ -4,21 +4,4 @@ const commonConfig = require('./webpack.common.config.js');
 
 module.exports = merge(commonConfig, {
   devtool: 'cheap-module-eval-source-map',
-  module: {
-    rules: [
-      {
-        test: /\.(ts|tsx)$/,
-        enforce: 'pre',
-        use: [
-          {
-            options: {
-              formatter: eslintFormatter,
-              eslintPath: require.resolve('eslint'),
-            },
-            loader: require.resolve('eslint-loader'),
-          },
-        ],
-      },
-    ],
-  },
 });