@Gregoor 2018-02-21
I'm still interested in landing this, but also afraid to put too much load on our servers. One way to solve that would be only server rendering when CPU usage < ~50% and else we'd fall back to client side rendering.