@Gregoor 2018-06-12
While I think we should revisit SSR at some point, we've moved to far apart from where this was branched off.